removed create-table-sql, added append flag

+++ plugins/operators/load_dimension.py
@@ -1,7 +1,6 @@
 from airflow.hooks.postgres_hook import PostgresHook
 from airflow.models import BaseOperator
 from airflow.utils.decorators import apply_defaults
-from airflow.contrib.hooks.aws_hook import AwsHook
 from plugins.helpers import sql_queries
 
 class LoadDimensionOperator(BaseOperator):
@@ -11,23 +10,22 @@ class LoadDimensionOperator(BaseOperator):
     def __init__(self,
                  redshift_conn_id="",
                  table="",
-                 create_table_sql="",
                  insert_sql="",
+                 append = "",
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id = redshift_conn_id
         self.table = table
-        self.create_table_sql = create_table_sql
         self.insert_sql = insert_sql
+        self.append = append
 
     def execute(self, context):
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
 
-        redshift.run(self.create_table_sql)
-
         self.log.info("Clearing data from destination Redshift table")
-        redshift.run("DELETE FROM {}".format(self.table))
+        if self.append == False:
+            redshift.run("DELETE FROM {}".format(self.table))
 
         self.log.info("Inserting data to songplays table")
         redshift.run(self.insert_sql)
